Attach the old cultist simulator experiment to zemn.me.
