--- project/cultist/multiplayer/BUILD.bazel
@@ -1,32 +1,0 @@
-load("//bzl:rules.bzl", "bazel_lint")
-load("//ts:rules.bzl", "ts_project")
-
-ts_project(
-    name = "index",
-    srcs = [
-        "index.tsx",
-    ],
-    resolve_json_module = True,
-    deps = [
-        "//:node_modules/@types/react",
-        "//:node_modules/@types/react-dom",
-        "//:node_modules/react",
-        "//:node_modules/react-dom",
-        "//project/cultist/multiplayer/pages",
-    ],
-)
-
-"""
-Maybe one day when I get CSS working again, I can remove this.
-web_app(
-    name = "multiplayer",
-    entry_points = ["//project/cultist/multiplayer:index.tsx"],
-    esbuild_deps = [":index"],
-    visibility = ["//project/cultist/multiplayer:__subpackages__"],
-    deps = [":index.html"],
-)
-
-
-"""
-
-bazel_lint(name = "bazel_lint")
