--- project/cultist/multiplayer/index.html
@@ -1,9 +1,0 @@
-<!DOCTYPE HTML>
-<head>
-	<title>Cultist Simulator Multiplayer</title>
-</head>
-<body>
-	<main>
-	</main>
-	<script type="module" src="index.js"></script>
-</body>
