--- project/cultist/multiplayer/index.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-
-import Home from '#root/project/cultist/multiplayer/pages/index.js';
-
-createRoot(document.querySelector('main')!).render(<Home />);
