--- project/cultist/multiplayer/pages/BUILD.bazel
@@ -1,22 +1,0 @@
-load("//bzl:rules.bzl", "bazel_lint")
-load("//ts:rules.bzl", "ts_project")
-
-package(default_visibility = ["//project/cultist/multiplayer:__subpackages__"])
-
-ts_project(
-    name = "pages",
-    srcs = [
-        "index.tsx",
-    ],
-    resolve_json_module = True,
-    deps = [
-        "//:node_modules/@types/react",
-        "//:node_modules/react",
-        "//project/cultist:save",
-        "//project/cultist/example",
-        "//project/cultist/react",
-        "//project/cultist/state",
-    ],
-)
-
-bazel_lint(name = "bazel_lint")
